gym_pybullet_drones/pyproject.toml
[tool.poetry]
name = "gym-pybullet-drones"
version = "0.0.3"
description = "PyBullet Gym environments for single and multi-agent reinforcement learning of quadcopter control"
authors = ["Jacopo Panerati <jacopo.panerati@utoronto.ca>"]
license = "MIT"
readme = "pypi_description.md"
repository = "https://github.com/utiasDSL/gym-pybullet-drones"
packages = [
    { include = "gym_pybullet_drones" }
]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "^1.22"
Pillow = "^9.0"
matplotlib = "^3.5"
cycler = "^0.10"
gym = "^0.21"
pybullet = "^3.2"
torch = "1.11.0"
"ray[rllib]" = "1.13"
stable-baselines3 = "1.5.0"
scipy = "^1.8"
tensorboard = "^2.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core @ git+https://github.com/python-poetry/poetry-core.git@master"]
build-backend = "poetry.core.masonry.api"
